% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_colors.R
\docType{data}
\name{cat_colors}
\alias{cat_colors}
\title{List of cat color palettes}
\format{An object of class \code{list} of length 4.}
\usage{
cat_colors
}
\description{
List of cat color palettes
}
\keyword{datasets}
