% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_palette.R
\name{cat_palette}
\alias{cat_palette}
\title{Generate palette for cat colors}
\usage{
cat_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of color (cat) palette}

\item{n}{Number of colors desired If omitted, uses all colours.}

\item{type}{Either "discrete" or "continuous",}
}
\description{
Generate palette for cat colors
}
\examples{
cat_palette(name="Pumpkin")
cat_palette(name="Chestnut", type="continuous")
cat_palette(name="SnowLion", n=100)

}
